requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
configuraptor==1.27.2
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
h11==0.16.0
hayhooks==0.7.0
haystack-ai==2.13.2
haystack-experimental==0.9.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lazy_imports==0.4.0
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.8.0
mcp-haystack==0.2.0
mdurl==0.1.2
more-itertools==10.7.0
msgspec==0.19.0
networkx==3.4.2
numpy==2.2.5
openai==1.78.0
packaging==25.0
plumbum==1.9.0
posthog==4.0.1
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
result==0.17.0
rich==14.0.0
rich-toolkit==0.14.5
rpds-py==0.24.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.4
starlette==0.46.2
tenacity==9.1.2
threadful==0.5.0
tomli==2.2.1
tomli_w==1.2.0
tqdm==4.67.1
typeguard==4.4.2
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uv==0.7.3
uvicorn==0.34.2
uvloop==0.21.0
uvx==1.0.0
watchfiles==1.0.5
websockets==15.0.1
